+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`1.11.1...main`][1.11.1...main].
 
+### Added
+
+- Added `Rules\Files\ReferenceNamespacedSymbolsRelativeToNamespacePrefixRector` to replace references to namespaced symbols (classes, functions, constants) whose fully-qualified name starts with a namespace prefix so they are relative to that prefix ([#295]), by [@localheinz]
+
 ## [`1.11.1`][1.11.1]
 
 For a full diff see [`1.11.0...1.11.1`][1.11.0...1.11.1].
@@ -247,5 +251,6 @@ For a full diff see [`fd198f0...0.1.0`][fd198f0...0.1.0].
 [#290]: https://github.com/ergebnis/rector-rules/pull/290
 [#291]: https://github.com/ergebnis/rector-rules/pull/291
 [#294]: https://github.com/ergebnis/rector-rules/pull/294
+[#295]: https://github.com/ergebnis/rector-rules/pull/295
 
 [@localheinz]: https://github.com/localheinz